Support custom fields in diffusion commit property views.
Summary:
This adds support for custom fields in diffusion property views.
Use case: Used by #wikimedia to parse the gerrit change-id field from
commits and display a link to gerrit in the commit property view.

Test Plan:
I tested as follows

1. Installed https://secure.phabricator.com/P1819
   (dropped the file into phabricator/extensions/)
2. viewed a commit with a Change-Id: field in the commit message.
3. Saw the Change-Id: in the property view, complete with a link to
   gerrit.

Note: P1819 is pretty messy but the code in this change should be
pretty solid. Some of it was lifted from the differential conduit api.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, qgil, Korvin

Differential Revision: https://secure.phabricator.com/D13581

Change-Id: I3ccf2ed556f9a6cb3d619ea68de704cc6e41bf8e