Prevent sending a header with unsafe characters [\r\n\0]
Summary:
This change prevents a malformed filename from breaking things.
Due to the code in AphrontHTTPSink->writeHeaders.

Test Plan:
If you attach a file to a maniphest task, and that file has a
newline or null in the filename, you get the following error
when attempting to download the file:

"Declining to emit response with unsafe HTTP header"

Reviewers: #blessed_reviewers, epriestley

Subscribers: Korvin

Differential Revision: https://secure.phabricator.com/D15843

Change-Id: If5f523bac22357f77679e6341388254e5adcc9eb