Ensure files live as long as intended.
Summary:
Previously, if a temporary file was created, but the same data was later intended to be permanent (e.g. a temporary project icon created for the Edit Picture page was assigned to a project for continued use), the existing temporary file would be retrieved by content hash, and then later cleaned up by the garbage collector undesirably. This change adjusts the file expiry time when retrieving exising files by content hash so they live as long as if the file did not exist but were newly created.

Fixes T10907.

Test Plan:
- Without the patch applied:
  - Assign a default picture to a project which is missing one
  - Navigate to the file and note that it is temporary (if it isn't, change the colour+icon combination of the project to one which hasn't been used before and repeat from the start)
- Apply the patch
- Assign the default picture to the project again
- Navigate to the file again and note that it is no longer temporary

Plus added a unit test. It is a little ugly due to how data types seem to work in LiskDAO.

Reviewers: epriestley, #blessed_reviewers!

Subscribers: cspeckmim, Korvin

Maniphest Tasks: T10907

Differential Revision: https://secure.phabricator.com/D16243