Fall back to parent tasks / subtasks when the task graph is big
Summary:
This avoids the performance implications of rendering very large
task graphs. This also allows parent tasks and subtasks to be
discoverable when the graph cannot be displayed efficiently.

Test Plan:
viewed a task with a large number of subtasks, saw the
old-fashioned list of subtasks instead of a graph.

Reviewers: #blessed_reviewers!, epriestley

Subscribers: Korvin

Differential Revision: https://secure.phabricator.com/D16304