WmfConfigSource: Attempt to load a readable config file in all cases
Maniphest Tasks: T146055, T151059

Summary:
Setting PHABRICATOR_ENV is fragile - there are a bunch of edge cases
where it isn't propogated to sub/remote processes.

Figuring out which file is readable and loading that as a fallback
actually makes sense because of the way we have divided privileges to
those config files - only one of the files should be owned by the same
gid as the active process so we use that one if the environement var
isn't set.

refs T146055
fixes T151059

Test Plan: tested locally for basic sanity. Unfortunately it is currently difficult to test this part of phabricator adequately.

Reviewers: chasemp, Krenair, Dzahn

Differential Revision: https://phabricator.wikimedia.org/D460